make bucket helper + finish nox
